+++ storm-core/src/jvm/backtype/storm/utils/ShellUtils.java
@@ -34,7 +34,7 @@ import org.slf4j.LoggerFactory;
 
 
 abstract public class ShellUtils {
-    public static Logger LOG = LoggerFactory.getLogger(ShellUtils.class);
+    public static final Logger LOG = LoggerFactory.getLogger(ShellUtils.class);
 
     // OSType detection
     public enum OSType {