+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 ## 0.11.0
+ * STORM-1143: Validate topology Configs during topology submission
+ * STORM-1157: Adding dynamic profiling for worker, restarting worker, jstack, heap dump, and profiling
+ * STORM-1123: TupleImpl - Unnecessary variable initialization.
+ * STORM-1153: Use static final instead of just static for class members.
+ * STORM-817: Kafka Wildcard Topic Support.
  * STORM-40: Turn worker garbage collection and heapdump on by default.
  * STORM-1152: Change map keySet iteration to entrySet iteration for efficiency.
  * STORM-1147: Storm JDBCBolt should add validation to ensure either insertQuery or table name is specified and not both.